Hooked PEBIL to instuctionAPI (an open-source instruction decoder from the dyninst project).
Implemented register liveness analysis to determine whether
flags need to be saved at each instrumentation point. Currently
we still use the same instrumentation point as before, but it
should be possible to better choose a point in each block where
flags are dead (if they aren't dead at the entry).



git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/PEBIL@3236 55925d41-9515-0410-9586-c6f35bcc715f