Added some documentation.
Fixed an issue occuring in cases where the stack was being patched up
in cases where the call was occuring in the entry block of a function.
This is normally ok, but we don't want it to occur when setSkipWrapper
has been set. This has been fixed, setSkipWrapper now implies stackIsSafe.

C MPI wrappers now call PMPI functions.